+++ bin/install.ps1
@@ -52,7 +52,7 @@ function main() {
                     Write-Host "Could not find an existing powershell modules directory to install to!" -ForegroundColor Red
                 }
             } else {
-                Write-Host "The pow command already exists in PowerShell!" -ForegroundColor Yellow
+                Write-Warning "The pow command already exists in PowerShell!"
                 Write-Host " Tip: Type install -force to re-install" -ForegroundColor Yellow
             }
             #if (Get-Command "pow" -errorAction SilentlyContinue) {$PowowShell=$true}