+++ bin/version.ps1
@@ -1,2 +1,2 @@
-"PowowShell v0.1.3"
+"PowowShell v0.1.4"
 "PowerShell v" + $PSVersionTable.PSVersion.toString()