+++ ide/js/step-grid.js
@@ -37,10 +37,14 @@ Vue.component('step-grid', {
             <tr v-for="row in rows">
                 <td v-for="step in row" :key="step.id" :id="step.id" :class="step.reference?'step drag':'step drag drop'">
                     <div :id="step.id" :class="'stepContainer'+(step.reference?' stepFilled':' stepEmpty')" @click="showDialog(step.id)">
-                        <div v-if="step.reference">
-                            <div class="stepName">{{ step.name }}</div>
-                            <div class="stepReference">{{step.reference}}</div>
-                        </div>
+                        <v-card height="200px" v-if="step.reference">
+                            <v-card-title class="blue white--text stepName" :title="step.name">
+                                {{ step.name }}
+                            </v-card-title>
+                            <v-card-text>
+                                <div class="stepReference">{{step.reference}}</div>
+                            </v-card-text>
+                        </v-card>
                     </div>
                 </td>
             </tr>