+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 CHANGELOG
 
+v0.1.4 24.03.2019
+* Use vuetify cards for steps
+* Load pipeline definition dynamically
+* Load components definition dynamically
+* Switched pow to UTF-8 output for components export
+* Removed path and extension from component reference
+* Esc to close step
+
 v0.1.3 23.03.2019
 * BUGFIX: Paths broken in pow build/run/verify etc
 * Dynamically load available inputs selector on step